requirements.txt
yfinance==0.1.63
mysql-connector-python==8.0.23
pandas==1.3.3
requests==2.26.0